Coroutines: make GameSessionManager load/save features 'suspend' functions.
All GameSessionManager functions that persist or retrieve game saves were synchronous and blocking. This change makes them 'suspend' functions, for use in coroutines. All such functions are dispatched in an injected ioDispatcher, with a save mutex handling access to save files and the cached save object.

GamePresenter, GameSetupPresenter, and MainActivity have been altered to use coroutines when accessing these functions.

This change also adds support for Kotlin Coroutines, adds a 'viewScope' property to BasePresenter, and updates the DI glue to inject Dispatchers.

Known Issue: use of coroutines in GameSetupPresenter.attach(..) results in visible misconfiguration of the View, with all sections initial visible before disappearing. This issue may be resolved once a full conversion to Coroutines for async operations is complete, but if not, this will need a fix.