Fix visual bug on GameSetup introduced by RxJava - > coroutines change.
The bug caused the complete layout (with all sections, even those not visible for the indicated Type) to be temporarily visible before being hidden; this was caused by moving a previously blocking game progress load to a coroutine.

The new change reworks how GameSetups are created by the Presenter, allowing one-time blocking view updates on creation with all subsequent updates occurring in a coroutine.