Refactor/coroutines (#3)
* Update Kotlin version to 1.9.22; necessary updates throughout.

* Update Gradle
* Update Hilt Plugin
* Target SDK 34
* JVM Target to 1.8
* Update Google and AndroidX libraries
* Add When clauses, non-null guarantees where necessary

* Coroutines: make GameSessionManager load/save features 'suspend' functions.

All GameSessionManager functions that persist or retrieve game saves were synchronous and blocking. This change makes them 'suspend' functions, for use in coroutines. All such functions are dispatched in an injected ioDispatcher, with a save mutex handling access to save files and the cached save object.

GamePresenter, GameSetupPresenter, and MainActivity have been altered to use coroutines when accessing these functions.

This change also adds support for Kotlin Coroutines, adds a 'viewScope' property to BasePresenter, and updates the DI glue to inject Dispatchers.

Known Issue: use of coroutines in GameSetupPresenter.attach(..) results in visible misconfiguration of the View, with all sections initial visible before disappearing. This issue may be resolved once a full conversion to Coroutines for async operations is complete, but if not, this will need a fix.

* Refactor: split GameSessionManager into Persistence/Creation managers. Shared package for (most) Game*Manager.

GamePersistenceManager now handles persistence of saved games, while GameCreationManager creates Game and related objects from persisted data or GameSetups.

* Refactor: GamePlayManager / GamePlaySession for running games.

Creates GamePlaySession, a coroutine-aware class for safely interacting with a Game and related structures (e.g. Evaluator and/or Solver); refactors GamePresenter to intertact only with a GamePlaySession rather than directly with Games.

* Refactor: VersionsManager now uses suspend functions. Updates FeatureAvailabilityPresenter to match.

* Refactor: GameRecordManager now uses "suspend" functions.

GameOutcomePresenter updated accordingly.

* Remove RxJava dependency, and all references to it.

The README still refers to RxJava as a tool previously used (replacing a promise to replace it with coroutines in the future).

* Fix visual bug on GameSetup introduced by RxJava - > coroutines change.

The bug caused the complete layout (with all sections, even those not visible for the indicated Type) to be temporarily visible before being hidden; this was caused by moving a previously blocking game progress load to a coroutine.

The new change reworks how GameSetups are created by the Presenter, allowing one-time blocking view updates on creation with all subsequent updates occurring in a coroutine.