Move all `domain` packages to `data`.
The Manager interfaces function more like data repositories than use cases. There are possible candidates here for UseCases, in particular 1. loading a persisted game based on seed/setup, and 2. creating a game play session from it, but conforming to any one use case implementation would require additional changes to the presentation classes to which they apply so it is left as another task.