Guess refactor: restructure Presentation DataModel "Guess" class.
This restructuring removes the explicit Constraint, allowing a subset of available data to be passed as the Guess and related classes (GuessLetter / GuessEvaluation). Currently, the BaseGuessAdapter and related Appearance classes are passed complete by-letter evaluations and selectively display only partial info depending on game type. By structuring Guess s.t. it supports partial data, responsibility for determining _how much_ information to display can be moved lower in the app layers -- to the Presenter or lower.

This change does not move Guess creation, only restructures the class to make such a move possible.