Guess refactor: move Guess instance creation to the GamePresenter.
Previously the GuessAdapter interface received Constraints and Strings, wrapping them in Guesses internally. Now the Presenter creates Guess instances and passes them up, with the GuessAdapter expecting inputs in that format. This does not change any aspect of the actual data being represented, only where the instances are made.

Guess instances are created in a few other places: e.g. within BaseGuessAdapter and related ViewHolders (as Placeholders for empty rows and pre-bind non-null data) and in GameInfoFragment (when formatting the How To Play example and legend).