Hints: better inference for #Included/#Exact feedback, based on "exclusivity positions."
In short, a character's exclusivity positions are positions in the word s.t. either the character appears there, _or_ the character does not appear anywhere. A character can be determined to be absent if its exclusivity positions cannot all be satisfied.

As an example, consider:

MIGHT (1 exact, 0 included)
SPITE (1 exact, 0 included)

"I" has exclusivity positions {1, 2} and "T" has exclusivity positions {3, 4}. If either set cannot be fully satisfied (e.g. because letters have at most 1 occurrence, because some other letter is known to fill one of those spaces, etc.,) then that letter is absent from the word. This inference can be performed when aggregated counts are known, but not when reduced to either Included or Exact counts.