Hint: draft of basic behavior.
Replace use of Constraint.MarkupType in Guess class (and related accessors) with a new GuessMarkup enumeration, that holds explicit values for "no markup applied" (equiv. null) and "acceptable in this position" (no equiv. in Constraint).

This change removes the final use of the Constraint class above the GamePresenter, and also allows hints to function legibly for Code Sequence games. Now, with Hints Off, the code colors are maintained except for characters explicitly marked NO. With Hints On, the code colors are universally replaced with the appropriate markup color in both the keyboard and game field.

There is also a change to hint behavior depending on whether English or Code secrets are used; see GameFeedbackManagerImpl. Specifically, Code-based games will never apply feedback other than NO or EMPTY when Hints are Off, while English-based games may apply ACCEPTABLE or potentially other feedbacks in those cases.