Feature/hints (#6)
* Guess refactor: restructure Presentation DataModel "Guess" class.

This restructuring removes the explicit Constraint, allowing a subset of available data to be passed as the Guess and related classes (GuessLetter / GuessEvaluation). Currently, the BaseGuessAdapter and related Appearance classes are passed complete by-letter evaluations and selectively display only partial info depending on game type. By structuring Guess s.t. it supports partial data, responsibility for determining _how much_ information to display can be moved lower in the app layers -- to the Presenter or lower.

This change does not move Guess creation, only restructures the class to make such a move possible.

* Guess refactor: move Guess instance creation to the GamePresenter.

Previously the GuessAdapter interface received Constraints and Strings, wrapping them in Guesses internally. Now the Presenter creates Guess instances and passes them up, with the GuessAdapter expecting inputs in that format. This does not change any aspect of the actual data being represented, only where the instances are made.

Guess instances are created in a few other places: e.g. within BaseGuessAdapter and related ViewHolders (as Placeholders for empty rows and pre-bind non-null data) and in GameInfoFragment (when formatting the How To Play example and legend).

* Hints: provide "hints" in the form of per-letter markup on Guesses. Partial implementation.

1. "Hint" system needs some internal revision, both to become more expressive and correct, and to limit and formalize behavior when "off".
2. Hints need a UI to be enabled / disabled.
3. Refactor desired for the FeedbackProvider interface to combine both Feedback types into a single object, and to allow callbacks / update flows.

* Hints: revised implementation for better flow-ability.

* Hints: properly display Code letters and How To Play in light of potential NO markup.

* Hints: minor fixes and cleanup

* Hints: interface draft for a system that allows hints to be activated manually.

* Hints: better inference for #Included/#Exact feedback, based on "exclusivity positions."

In short, a character's exclusivity positions are positions in the word s.t. either the character appears there, _or_ the character does not appear anywhere. A character can be determined to be absent if its exclusivity positions cannot all be satisfied.

As an example, consider:

MIGHT (1 exact, 0 included)
SPITE (1 exact, 0 included)

"I" has exclusivity positions {1, 2} and "T" has exclusivity positions {3, 4}. If either set cannot be fully satisfied (e.g. because letters have at most 1 occurrence, because some other letter is known to fill one of those spaces, etc.,) then that letter is absent from the word. This inference can be performed when aggregated counts are known, but not when reduced to either Included or Exact counts.

* Hint: revise activation button behavior.

Button now shows Action, not State
Button is hidden for games that do not support Hints
Button is disabled until Hints are available (Markup would change somewhere)
Messages displayed incorrect attempt to activate hints and upon first hint activation.

* Hint: draft of basic behavior.

Replace use of Constraint.MarkupType in Guess class (and related accessors) with a new GuessMarkup enumeration, that holds explicit values for "no markup applied" (equiv. null) and "acceptable in this position" (no equiv. in Constraint).

This change removes the final use of the Constraint class above the GamePresenter, and also allows hints to function legibly for Code Sequence games. Now, with Hints Off, the code colors are maintained except for characters explicitly marked NO. With Hints On, the code colors are universally replaced with the appropriate markup color in both the keyboard and game field.

There is also a change to hint behavior depending on whether English or Code secrets are used; see GameFeedbackManagerImpl. Specifically, Code-based games will never apply feedback other than NO or EMPTY when Hints are Off, while English-based games may apply ACCEPTABLE or potentially other feedbacks in those cases.