Cleanup: regenerate (nearly) all icons to Material Symbols.
The only remaining legacy icons are the custom-made PNGs for Puzzle Info and Create Puzzle. These icons are based on the "grid view" material icon, with the bottom-right grid item replaced by an Info "i" or Create "+" respectively. Updating these icons requires working with rasterized or custom SVGs of the source icon(s).