Fix/icons (#8)
* Cleanup: regenerate (nearly) all icons to Material Symbols.

The only remaining legacy icons are the custom-made PNGs for Puzzle Info and Create Puzzle. These icons are based on the "grid view" material icon, with the bottom-right grid item replaced by an Info "i" or Create "+" respectively. Updating these icons requires working with rasterized or custom SVGs of the source icon(s).

* Cleanup: updates Puzzle Info and Puzzle Create icons for consistency with new Material Symbols theme.

* Cleanup: update documentation (NOTICE and credits) to refer to the new Material Symbols.

* Icons: add mdpi versions of the two custom icons, left out of previous commits.