Feature/disable update check (#20)
* Allow "Check for updates" to be disabled.

If disabled, no network calls will be made. The Daily will remain available, with a warning that it is local-only and may not match those of other people.

Internally, this is implemented by a change to the "Launch" process from game setup: the GameSetup for the Daily is set up as before, but "daily = false" is set on the version sent to launch. The puzzle itself is thus identical but everywhere outside of setup it is treated as a seeded, not Daily, challenge.

* Fix: incorrect display of "Checking for availabilitiy...." when update check is off.

* Warn users who enter a Local Daily seed that it may lock them out of the real Daily later.

* Organize HW Keyboard and Update Check settings under "Advanced Settings" category.

* Fix layout of SettingsActivity