Feature/usability improvements (#34)
* Custom button to start puzzle in Game Setup activity topbar.

Previously the "Create Puzzle" button was a nondescript icon in the top bar; some users reported trouble identifying the interaction. A custom view is now used to display a large, colorful button in the tapbar named "Create".

* Add "Copy" button to Edit Seed dialog.

Seed value can already be copied using normal EditText interactions, but this streamlines the process and emphasizes that the interaction is possible and likely desired.

The button is available on the Edit dialog to avoid icon clutter in the main setup screen.

* Consolidate "valid word dictionary" into an extended document.

Previously "valid" lists were length-specific and included some gaps when compared to similar word lists, e.g. the Scrabble dictionary. The new consolidated file has words of all lengths (no delays noticed for the larger file) and is more comprehensive.

Note: this word list is used exclusively for determining whether a word entered by the user is a valid English word. It is NEVER used to generate secrets or automatic guesses/hints.